[REFACTOR] #53 : domain model 패키지 정리 및 네이밍 수정
